Replace the fs2 crate which is no longer maintained with fs4 and rustix
This change originally started with some qdrant test failures on FreeBSD/ZFS
because posix_fallocate returns `EINVAL` when attempting to create a sparse file
on a copy-on-write filesystem. While investigating I found that this repository
relied on fs2 which is no longer maintained. There is an
[fs4](https://github.com/al8n/fs4-rs) crate which
continues development, but is working to replace use of the `libc` crate with
`rustix`.

Since fs4 depends on rustix anyways, there's no additional linkage needed to
pull some symbols directly from rustix for use on FreeBSD.

This should be the best of both worlds and hopefully this library can move away
from depending on fs4 as time goes on